python/requirements.txt
backcall==0.2.0
debugpy==1.4.1
decorator==5.0.9
ipykernel==6.1.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
numpy==1.21.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.10.0
pyserial==3.5
python-dateutil==2.8.2
pyzmq==22.2.1
six==1.16.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
sovol_xy
